+++ packages/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @next-ai-ready/next
 
+## 0.1.0-alpha.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @next-ai-ready/mcp@0.1.0-alpha.16
+
 ## 0.1.0-alpha.18
 
 ### Patch Changes